fixed bug

+++ common/util.go
@@ -121,8 +121,8 @@ func CalculateWeeklyAndMonthlyActivationData(dailyTransactionsCountList []Activa
 	return weeklyTransactionsCount, monthlyTransactionsCount
 }
 
-func GetDateByTimeStamp(timestamp uint64) string {
-	t := time.Unix(int64(timestamp), 0)
+func GetDateByTimeStamp(timestamp int64) string {
+	t := time.Unix(timestamp, 0)
 	date := fmt.Sprintf("%04d_%02d_%02d", t.Year(), int(t.Month()), t.Day())
 	return date
 }
